Add .travis.yml
Build and test the repository against stable, beta and nightly.
Run rustfmt and clippy, but only on nightly because clippy only
supports nightly and we don't want different versions of rustfmt
checking the same code. Sometimes clippy takes time to catch
up with nightly breaking changes, so the build won't fail if
clippy fails to install.